+++ src/r3f/components/CanvasDOMOverlay.d.ts
@@ -1,0 +1,5 @@
+import type { ComponentPropsWithoutRef, ReactNode } from 'react';
+
+export interface CanvasDOMOverlayProps extends ComponentPropsWithoutRef<'div'> {
+    children?: ReactNode;
+}