+++ src/r3f/index.d.ts
@@ -1,0 +1,7 @@
+export * from './components/CameraControls';
+export * from './components/CameraTransition';
+export * from './components/CanvasDOMOverlay';
+export * from './components/CompassGizmo';
+export * from './components/SettledObjects';
+export * from './components/TilesAttributionOverlay';
+export * from './components/TilesRenderer';